# docs/openapi-public.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NexPlay Public API",
    "version": "1.0.0",
    "description": "API publique en lecture pour intégrations et futurs plugins. Authentification via header X-Api-Key (créer une clé dans /api/developer/keys)."
  },
  "servers": [{ "url": "http://localhost:4000/public" }],
  "security": [{ "ApiKeyAuth": [] }],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    }
  },
  "paths": {
...
    "/v1/leaderboards/{gameId}": {
      "get": {
        "summary": "Classement par jeu",
        "parameters": [
          { "name": "gameId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "scope", "in": "query", "schema": { "type": "string", "enum": ["world", "continent", "country", "region"] } },
          { "name": "continent", "in": "query", "schema": { "type": "string" } },
          { "name": "country", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Entries classées" } }
      }
    },
...
  }
}